requirements.txt
docopt
spacy
transformers
torch 
torchvision
fairseq
sacremoses
flask 
python-dotenv
google-cloud-translate
more_itertools
pytest
simalign